pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = ">=3.10.0,<3.12"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
openai = "^1.38.0"
beautifulsoup4 = "^4.12.3"
fake-useragent = "^1.5.1"
tiktoken = "^0.7.0"
praw = "^7.7.1"
replit = "^4.0.0"
scrapy = "^2.11.2"
twisted = "^24.7.0"
pandas = "^2.2.2"
playwright = "^1.45.1"
twikit = "^2.1.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"